+++ app/src/main/java/site/leos/apps/lespas/album/AlbumDetailFragment.kt
@@ -175,7 +175,7 @@ class AlbumDetailFragment : Fragment(), ActionMode.Callback {
                 }
             }
         }
-        context?.registerReceiver(snapseedCatcher, IntentFilter(CHOOSER_SPY_ACTION))
+        requireContext().registerReceiver(snapseedCatcher, IntentFilter(CHOOSER_SPY_ACTION))
 
         // Content observer looking for Snapseed output
         snapseedOutputObserver = object : ContentObserver(Handler(Looper.getMainLooper())) {
@@ -273,8 +273,8 @@ class AlbumDetailFragment : Fragment(), ActionMode.Callback {
                 PhotoGridAdapter.PhotoDetailsLookup(this),
                 StorageStrategy.createStringStorage()
             ).withSelectionPredicate(object : SelectionTracker.SelectionPredicate<String>() {
-                override fun canSetStateForKey(key: String, nextState: Boolean): Boolean = (key.isNotEmpty())
-                override fun canSetStateAtPosition(position: Int, nextState: Boolean): Boolean = (position != 0)
+                override fun canSetStateForKey(key: String, nextState: Boolean): Boolean = key.isNotEmpty()
+                override fun canSetStateAtPosition(position: Int, nextState: Boolean): Boolean = position > 0
                 override fun canSelectMultiple(): Boolean = true
             }).build().apply {
                 addObserver(object : SelectionTracker.SelectionObserver<String>() {
@@ -925,7 +925,7 @@ class AlbumDetailFragment : Fragment(), ActionMode.Callback {
         internal fun getPhotoPosition(photoId: String): Int = currentList.indexOfLast { it.id == photoId }
         class PhotoKeyProvider(private val adapter: PhotoGridAdapter): ItemKeyProvider<String>(SCOPE_CACHED) {
             override fun getKey(position: Int): String = adapter.getPhotoId(position)
-            override fun getPosition(key: String): Int = with(adapter.getPhotoPosition(key)) { if (this >= 0) this else RecyclerView.NO_POSITION }
+            override fun getPosition(key: String): Int = adapter.getPhotoPosition(key)
         }
         class PhotoDetailsLookup(private val recyclerView: RecyclerView): ItemDetailsLookup<String>() {
             override fun getItemDetails(e: MotionEvent): ItemDetails<String>? {